+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2023-06-29: added `check-meta` command to CLI.
 - 2023-06-28: updated packages.
 - 2023-04-13:
   - fixes in default query in CLI tool build-sql command.